requirements.txt
requests==2.28.2
google-auth==2.16.0
google-auth-oauthlib==0.5.2
google-auth-httplib2==0.1.0
google-api-python-client==2.91.0
python-dotenv==0.21.0  # Only for local testing if using .env
